pyproject.toml
[project]
name = "merlot"
version = "0.1.0"
description = ""
authors = [
    {name = "V. Guilherme",email = "v.guilherme.barreto@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-alembic (>=3.1.1,<4.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.11.3,<0.12.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-cov (>=6.1.0,<7.0.0)",
    "twilio (>=9.5.2,<10.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.ruff]
line-length = 90
    exclude = [
    '.bzr',
    '.direnv',
    '.eggs',
    '.git',
    '.git-rewrite',
    '.hg',
    '.ipynb_checkpoints',
    '.mypy_cache',
    '.nox',
    '.pants.d',
    '.pyenv',
    '.pytest_cache',
    '.pytype',
    '.ruff_cache',
    '.svn',
    '.tox',
    '.venv',
    '.vscode',
    '__pypackages__',
    '_build',
    'buck-out',
    'build',
    'dist',
    'migrations',
    'node_modules',
    'site-packages',
    'venv'
]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'python merlot/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=merlot -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'